New staggered isometric mapping function
Previous functon found the drawing bounds of a tile for the given
point, but could not differenciate actual tiles.

This is now solved via algebra.

Moved isoadventure bounds checks to the bottom of the sprite to align
his floor tile coordinate.

Boungs checking still needs work, and walkable testing by property is
still not working.